вернул константы в модули и удалил setup.js; модуль filter функция checkFeatures заменил второй if на тернарный оператор; функция checkFilters рефакторинг; модуль form добавил сброс валидатора при нажатии на кнопку reset
